refactor(mm-next): rename config.js to config.mjs
