+++ packages/mirror-media-next/apollo/apollo-client.js
@@ -1,6 +1,6 @@
 import { ApolloClient, InMemoryCache } from '@apollo/client'
 
-import { API_HOST } from '../config'
+import { API_HOST } from '../config/index.mjs'
 
 const client = new ApolloClient({
   uri: `https://${API_HOST}/api/graphql`,
