+++ packages/mirror-media-next/config/index.mjs
