+++ packages/mirror-media-next/pages/author/[id].js
@@ -5,7 +5,7 @@ import client from '../../apollo/apollo-client'
 import { fetchContact } from '../../apollo/query/contact'
 import { fetchPosts } from '../../apollo/query/posts'
 import AuthorArticles from '../../components/author-articles'
-import { GCP_PROJECT_ID } from '../../config'
+import { GCP_PROJECT_ID } from '../../config/index.mjs'
 
 const AuthorContainer = styled.main`
   width: 320px;
